refactor structs
