+++ src/structs/mod.rs
@@ -1,0 +1,5 @@
+mod entry;
+mod filetype;
+mod icons;
+
+pub use crate::structs::{entry::Entry, filetype::Filetype, icons::Icon, icons::Icons};
